[unfinished - part 1] partially implement refreshing views with jQuery Ajax

+++ DepositDatabase/CurrenciesData.cs
@@ -7,7 +7,7 @@ namespace DepositDatabase
 {
     public class CurrenciesData
     {
-        public static Currency GetCurrencyById(byte id)
+        public static Currency GetCurrencyById(int id)
         {
             return DepositDbContextExtension.GetInstance().Currencies.Find(id);
         }
