+++ DepositDatabase/Handlers/NewDepositHandler.cs
@@ -11,12 +11,12 @@ namespace DepositDatabase.Handlers
             return CardsData.GetCardById(id).ToDomainLogic();
         }
 
-        public DomainLogic.Model.Currency GetCurrencyByDepositTermsId(byte termsId)
+        public DomainLogic.Model.Currency GetCurrencyByDepositTermsId(int termsId)
         {
             return DepositTermsData.GetTermById(termsId).Currency.ToDomainLogic();
         }
 
-        public DomainLogic.Model.Deposit CreateNewDeposit(decimal depositAmount, byte wayOfAccumulationId, string userId, byte termsId, string cardId)
+        public DomainLogic.Model.Deposit CreateNewDeposit(decimal depositAmount, int wayOfAccumulationId, string userId, int termsId, string cardId)
         {
             var depositTerm = DepositTermsData.GetTermById(termsId);
