+++ DomainLogic/Handlers/ICloseDepositHandler.cs
@@ -4,7 +4,7 @@ namespace DomainLogic.Handlers
 {
     public interface ICloseDepositHandler
     {
-        Currency GetCurrencyByDepositTermsId(byte depositTermsId);
+        Currency GetCurrencyByDepositTermsId(int depositTermsId);
         Deposit GetDepositById(int id);
         void IncreaseCardBalanceByDepositBalance(decimal depositBalance, string cardId);
         void SetDepositState(string name, int depositId);
